requirements.txt
fastmcp
pyjwt
pytest
google-genai>=0.15.0
requests>=2.31.0
